+++ addOns/dev/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
+### Added
+- Extra protected pages to simple-json-cookie to ensure spidering really works.
+
+### Fixed
+- Issue where folder level pages without a trailing slash did not link correctly to sub pages.
 
 
 ## [0.6.0] - 2024-07-22